Add VWProjectRisksTable components and implement role-based access control for editing and deleting risks
