+++ Clients/src/presentation/vw-v2-components/Table/VWProjectRisksTableHead.tsx
@@ -1,0 +1,37 @@
+import { TableCell, TableHead, TableRow } from "@mui/material";
+import singleTheme from "../../themes/v1SingleTheme";
+
+
+const VWProjectRisksTableHead = ({ columns }: { columns: any[] }) => {
+  return (
+    <TableHead
+      sx={{
+        backgroundColor:
+          singleTheme.tableStyles.primary.header.backgroundColors,
+      }}
+    >
+      <TableRow sx={singleTheme.tableStyles.primary.header.row}>
+        {columns.map((column, index) => (
+          <TableCell
+            key={index}
+            style={{
+              ...singleTheme.tableStyles.primary.header.cell,
+              ...(index === columns.length - 1
+                ? {
+                  position: "sticky",
+                  right: 0,
+                  backgroundColor:
+                    singleTheme.tableStyles.primary.header.backgroundColors,
+                }
+                : {}),
+            }}
+          >
+            {column}
+          </TableCell>
+        ))}
+      </TableRow>
+    </TableHead>
+  );
+};
+
+export default VWProjectRisksTableHead;
