+++ Clients/src/presentation/vw-v2-components/Table/index.tsx
@@ -3,232 +3,20 @@ import {
   Table,
   TableBody,
   TableCell,
-  TableContainer,
-  TableHead,
+  TableContainer, 
   TablePagination,
   TableRow,
   Typography,
   useTheme,
-  Box,
 } from "@mui/material";
 import singleTheme from "../../themes/v1SingleTheme";
-import { useCallback, useContext, useMemo, useState } from "react";
+import { useCallback, useMemo, useState } from "react";
 import TablePaginationActions from "../../components/TablePagination";
 import { ReactComponent as SelectorVertical } from "../../assets/icons/selector-vertical.svg";
-import { ProjectRisk } from "../../../domain/types/ProjectRisk";
-import { VerifyWiseContext } from "../../../application/contexts/VerifyWise.context";
-import { RISK_LABELS } from "../../components/RiskLevel/constants";
-import IconButton from "../../components/IconButton";
 import placeholderImage from "../../assets/imgs/empty-state.svg";
-import { formatDate } from "../../tools/isoDateToString";
-import allowedRoles from "../../../application/constants/permissions";
+import VWProjectRisksTableHead from "./VWProjectRisksTableHead";
+import VWProjectRisksTableBody from "./VWProjectRisksTableBody";
 
-const VWProjectRisksTableHead = ({ columns }: { columns: any[] }) => {
-  return (
-    <TableHead
-      sx={{
-        backgroundColor:
-          singleTheme.tableStyles.primary.header.backgroundColors,
-      }}
-    >
-      <TableRow sx={singleTheme.tableStyles.primary.header.row}>
-        {columns.map((column, index) => (
-          <TableCell
-            key={index}
-            style={{
-              ...singleTheme.tableStyles.primary.header.cell,
-              ...(index === columns.length - 1
-                ? {
-                  position: "sticky",
-                  right: 0,
-                  backgroundColor:
-                    singleTheme.tableStyles.primary.header.backgroundColors,
-                }
-                : {}),
-            }}
-          >
-            {column}
-          </TableCell>
-        ))}
-      </TableRow>
-    </TableHead>
-  );
-};
-
-const VWProjectRisksTableBody = ({
-  rows,
-  page,
-  rowsPerPage,
-  setSelectedRow,
-  setAnchor,
-  onDeleteRisk,
-  flashRow,
-}: {
-  rows: any[];
-  page: number;
-  rowsPerPage: number;
-  setSelectedRow: any;
-  setAnchor: any;
-  onDeleteRisk: (id: number) => void;
-  flashRow: number | null;
-}) => {
-  const { setInputValues, users, userRoleName } = useContext(VerifyWiseContext);
-  const cellStyle = singleTheme.tableStyles.primary.body.cell;
-  const theme = useTheme();
-  const handleEditRisk = (row: any, event?: React.SyntheticEvent) => {
-    if (allowedRoles.projectRisks.edit.includes(userRoleName)) {
-      setSelectedRow(row);
-      setInputValues(row);
-      setAnchor(event?.currentTarget);
-    }
-  };
-
-  const handleDeleteRisk = async (riskId: number) => {
-    onDeleteRisk(riskId);
-  };
-
-  const displayUserFullName = (userId: number) => {
-    const currentUser = users.find(
-      (user: any) => user.id === userId
-    );
-    const fullName = currentUser
-      ? `${currentUser.name} ${currentUser.surname}`
-      : "";
-    return fullName.length > 30 ? `${fullName.slice(0, 30)}...` : fullName;
-  };
-
-  return (
-    <TableBody>
-      {rows &&
-        rows
-          .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-          .map((row: ProjectRisk, index: number) => (
-            <TableRow
-              key={index}
-              sx={singleTheme.tableStyles.primary.body.row}
-              onClick={(e) => handleEditRisk(row, e)}
-              style={{
-                cursor: allowedRoles.projectRisks.edit.includes(userRoleName)
-                  ? "pointer"
-                  : "default",
-              }}
-            >
-              <TableCell
-                sx={cellStyle}
-                style={{
-                  backgroundColor: flashRow === row.id ? "#e3f5e6" : "",
-                }}
-              >
-                {row.risk_name
-                  ? row.risk_name?.length > 30
-                    ? `${row.risk_name.slice(0, 30)}...`
-                    : row.risk_name
-                  : "-"}
-              </TableCell>
-              <TableCell
-                sx={cellStyle}
-                style={{
-                  backgroundColor: flashRow === row.id ? "#e3f5e6" : "",
-                }}
-              >
-                {row.risk_owner ? displayUserFullName(Number(row.risk_owner)) : "-"}
-              </TableCell>
-              <TableCell
-                sx={cellStyle}
-                style={{
-                  backgroundColor: flashRow === row.id ? "#e3f5e6" : "",
-                }}
-              >
-                {row.severity ? row.severity : "-"}
-              </TableCell>
-              <TableCell
-                sx={cellStyle}
-                style={{
-                  backgroundColor: flashRow === row.id ? "#e3f5e6" : "",
-                }}
-              >
-                {row.likelihood ? row.likelihood : "-"}
-              </TableCell>
-              <TableCell
-                sx={cellStyle}
-                style={{
-                  backgroundColor: flashRow === row.id ? "#e3f5e6" : "",
-                }}
-              >
-                {row.mitigation_plan
-                  ? row.mitigation_plan.length > 30
-                    ? `${row.mitigation_plan.slice(0, 30)}...`
-                    : row.mitigation_plan
-                  : "-"}
-              </TableCell>
-              <TableCell
-                sx={cellStyle}
-                style={{
-                  backgroundColor: flashRow === row.id ? "#e3f5e6" : "",
-                }}
-              >
-                {row.mitigation_status ? row.mitigation_status : "-"}
-              </TableCell>
-              <TableCell
-                sx={cellStyle}
-                style={{
-                  backgroundColor: flashRow === row.id ? "#e3f5e6" : "",
-                }}
-              >
-                <Box
-                  sx={{
-                    backgroundColor:
-                      Object.values(RISK_LABELS).find(
-                        (risk) => risk.text === row.risk_level_autocalculated
-                      )?.color || "transparent",
-                    borderRadius: theme.shape.borderRadius,
-                    padding: "8px",
-                    textAlign: "center",
-                    justifyContent: "center",
-                    color: "white",
-                  }}
-                >
-                  {row.risk_level_autocalculated
-                    ? row.risk_level_autocalculated
-                    : "-"}
-                </Box>
-              </TableCell>
-              <TableCell
-                sx={cellStyle}
-                style={{
-                  backgroundColor: flashRow === row.id ? "#e3f5e6" : "",
-                }}
-              >
-                {row.deadline ? formatDate(row.deadline.toString()) : "NA"}
-              </TableCell>
-              <TableCell
-                sx={{
-                  ...singleTheme.tableStyles.primary.body.cell,
-                  position: "sticky",
-                  right: 0,
-                  minWidth: "50px",
-                }}
-                style={{
-                  backgroundColor: flashRow === row.id ? "#e3f5e6" : "",
-                }}
-              >
-                {allowedRoles.projectRisks.edit.includes(userRoleName) &&
-                  <IconButton
-                    id={row.id}
-                    type="risk"
-                    onMouseEvent={(e) => handleEditRisk(row, e)}
-                    onDelete={() => handleDeleteRisk(row.id)}
-                    onEdit={() => handleEditRisk(row)}
-                    warningTitle="Delete this project risk?"
-                    warningMessage="Are you sure you want to delete this project risk. This action is non-recoverable."                  
-                  />
-                }
-              </TableCell>
-            </TableRow>
-          ))}
-    </TableBody>
-  );
-};
 
 const VWProjectRisksTable = ({
   columns,
